Remove ntap-core
