--- .gitattributes
@@ -1,2 +1,0 @@
-*.css linguist-vendored
-ntap-desktop/public/* linguist-vendored
