+++ README.md
@@ -1,28 +1,24 @@
-<div align="center">
-    <img src="resources/ntap-logo-128x128.png" alt="ntap - Real-time network utilization monitoring tool"><br>
-    <h1>ntap</h1>
-    <p>
-        <strong>Real-time network utilization monitoring tool.</strong>
-    </p>
-</div>
+[crates-badge]: https://img.shields.io/crates/v/ntap.svg
+[crates-url]: https://crates.io/crates/ntap
+[license-badge]: https://img.shields.io/crates/l/ntap.svg
+
+# ntap [![Crates.io][crates-badge]][crates-url] ![License][license-badge]
 
 ## Overview
 **ntap** is a real-time network utilization monitoring tool.  
 It provides comprehensive insights into your network's activity, enabling users to monitor traffic, manage connections, and view network configurations with ease.
 
+![image](resources/ntap-ss-overview.png)  
+
 ## Features
 - **Real-time Monitoring**: Track network utilization as it happens.
 - **Connection Management**: Analyze active connections quickly and effectively.
 - **Interface and Routing Insights**: Get detailed views of network interfaces and routing tables.
-- **Two versions available**: CLI app and Desktop app are available, catering to different user preferences.
 
 ## Prerequisites
 - Ensure you have a compatible operating system (Linux, macOS, Windows).
 
 ## Installation
-
-## CLI: ntap
-![image](resources/ntap-ss-overview.png)  
 
 ### Install prebuilt binaries via shell script
 
@@ -68,24 +64,6 @@ cargo build --release
 ./target/release/ntap
 ```
 
-## Desktop Application: ntap-desktop
-![image](resources/ntap-desktop-ss-overview.png)
-
-#### Using Installer
-Download the installer for your platform from the [releases page](https://github.com/shellrow/ntap/releases).
-
-#### Build from source
-First, clone the repository:
-```
-git clone https://github.com/shellrow/ntap
-```
-Then, build the project (assuming Rust and Tauri are already installed):
-```
-cd ntap-desktop
-cargo tauri build
-```
-Run the installer in the dist directory.
-
 ## Post-Install Configuration
 
 The following post-install configuration steps are applicable to both the CLI version (`ntap`) and the desktop application (`ntap-desktop`).  
@@ -100,12 +78,12 @@ These steps ensure that `ntap` has the necessary permissions and environment set
 Granting capabilities to the `ntap` binary allows it to operate with the necessary privileges without requiring `sudo` for each execution.  
 This method is recommended for single-user machines or in environments where all users are trusted.
 
-Assign necessary capabilities to the ntap binary
+Assign necessary capabilities to the `ntap` binary
 ```sh
 sudo setcap 'cap_sys_ptrace,cap_dac_read_search,cap_net_raw,cap_net_admin+ep' $(command -v ntap)
 ```
 
-Run ntap as an unprivileged user:
+Run `ntap` as an unprivileged user:
 ```sh
 ntap
 ```
@@ -115,7 +93,7 @@ ntap
 - `cap_net_raw,cap_net_admin`: Enables packet capturing capabilities.
 
 #### 2. Using `sudo` (for multi-user environments)
-For environments with multiple users, requiring privilege escalation each time ntap is run can enhance security.
+For environments with multiple users, requiring privilege escalation each time `ntap` is run can enhance security.
 ```
 sudo ntap
 ```
@@ -144,18 +122,9 @@ chmod-bpf check
 sudo chmod-bpf install
 ```
 
-#### macOS Security
-After installing Ntap Desktop on macOS, you may encounter a security warning that prevents the app from opening, stating that it is from an unidentified developer. This is a common macOS security measure for apps downloaded outside of the App Store.
-
-To resolve this issue and open Ntap Desktop, you can remove the security attributes that macOS assigns to the application using the following command in the Terminal:
-
-```sh
-xattr -rc "/Applications/Ntap Desktop.app"
-```
-
 ### Post-Install (Windows)
 - Ensure that you have [Npcap](https://npcap.com/#download) installed, which is necessary for packet capturing on Windows
 - Download and install Npcap from [Npcap](https://npcap.com/#download). Choose the "Install Npcap in WinPcap API-compatible Mode" during installation.
 
 ### License
-ntap is released under the MIT License. See the LICENSE file for more details.
+`ntap` is released under the MIT License. See the LICENSE file for more details.
