--- ntap-core/README.md
@@ -1,2 +1,0 @@
-# ntap-core
-Core library for ntap and ntap-desktop
