--- ntap-core/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod db;
-pub mod deps;
-pub mod net;
-pub mod sys;
-pub mod config;
-pub mod log;
-pub mod notification;
-pub mod process;
-pub mod time;
