+++ ntap/Cargo.toml
@@ -7,7 +7,7 @@ description = "Real-time network utilization monitoring tool"
 repository = "https://github.com/shellrow/ntap"
 homepage = "https://github.com/shellrow/ntap"
 documentation = "https://github.com/shellrow/ntap"
-readme = "README.md"
+readme = "../README.md"
 keywords = ["network","security","cross-platform"]
 categories = ["network-programming"]
 license = "MIT"
@@ -18,6 +18,7 @@ serde_json = "1"
 log = "0.4"
 simplelog = "0.12"
 netdev = { version = "0.28", features = ["serde"] }
+nex = { version = "0.15", features = ["serde"] }
 tokio = { version = "1.35" }
 clap = { version = "4.4", features = ["cargo"] }
 crossterm = "0.27"
@@ -28,8 +29,31 @@ color-eyre = "0.6"
 palette = "0.7"
 ratatui = "0.25"
 comfy-table = "7.1"
-ntap-core = { path = "../ntap-core", version = "0.3.0" }
+hickory-resolver = { version = "0.24" }
+futures = {version = "0.3"}
+netsock = { version = "0.2", features = ["serde"] }
+reqwest = { version="0.12", default-features = false, features = ["json", "rustls-tls", "stream"] }
+chrono = { version = "0.4", features = ["serde"] }
+time = { version = "0.3", features = ["local-offset"] }
+ipnet = "2.5"
+ipstruct = "0.2"
+home = "0.5"
+bincode = "1.3"
+rangemap = "1.4"
+ntap-db-as = { path = "../db/ntap-db-as", version = "0.3.0" }
+ntap-db-country = { path = "../db/ntap-db-country", version = "0.3.0" }
+ntap-db-ipv4-asn = { path = "../db/ntap-db-ipv4-asn", version = "0.3.0" }
+ntap-db-ipv4-country = { path = "../db/ntap-db-ipv4-country", version = "0.3.0" }
+ntap-db-ipv6-asn = { path = "../db/ntap-db-ipv6-asn", version = "0.3.0" }
+ntap-db-ipv6-country = { path = "../db/ntap-db-ipv6-country", version = "0.3.0" }
+ntap-db-oui = { path = "../db/ntap-db-oui", version = "0.3.0" }
+ntap-db-tcp-service = { path = "../db/ntap-db-tcp-service", version = "0.3.0" }
+ntap-db-udp-service = { path = "../db/ntap-db-udp-service", version = "0.3.0" }
 
 [target.'cfg(windows)'.dependencies]
 indicatif = "0.16"
 inquire = "0.6"
+winreg = "0.50"
+sha2 = "0.10"
+zip = "0.6"
+privilege = "0.3"
