+++ ntap/src/app.rs
@@ -1,6 +1,8 @@
+#![allow(unused)]
+
 use std::time::Duration;
 
-use ntap_core::{
+use crate::{
     config::AppConfig,
     net::{host::HostDisplayInfo, service::ServiceDisplayInfo, stat::NetStatData, socket::SocketDisplayInfo},
     process::ProcessDisplayInfo,
