+++ ntap/src/db/mod.rs
@@ -1,2 +1,4 @@
+#![allow(unused)]
+
 pub mod ip;
 pub mod service;
