+++ ntap/src/deps/mod.rs
@@ -1,3 +1,5 @@
+#![allow(unused)]
+
 #[cfg(not(target_os = "windows"))]
 mod unix;
 #[cfg(not(target_os = "windows"))]
